+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Open the `UploadedFile` object in `#to_rack_response` in `rack_response` plugin (@janko)
+
 * Fix `store_dimensions` plugin making second argument in `Shrine#extract_metadata` mandatory (@jrochkind)
 
 ## 2.15.0 (2019-02-08)